+++ packages/wasm-dalek-ed25519/test/wasm.js
@@ -1,0 +1,11 @@
+// Copyright 2019 @polkadot/wasm-schnorrkel authors & contributors
+// This software may be modified and distributed under the terms
+// of the Apache-2.0 license. See the LICENSE file for details.
+// @ts-check
+
+const { beforeAll, runAll } = require('./all');
+
+(async () => {
+  await beforeAll();
+  runAll();
+})();