+++ packages/wasm-schnorrkel/src/lib.rs
@@ -6,10 +6,6 @@
 // which was adpated from the initial https://github.com/paritytech/schnorrkel-js/
 // forked at commit eff430ddc3090f56317c80654208b8298ef7ab3f
 
-extern crate wasm_bindgen;
-extern crate wee_alloc;
-extern crate schnorrkel;
-
 use schnorrkel::{
 	Keypair, MiniSecretKey, PublicKey, SecretKey, Signature,
 	derive::{Derivation, ChainCode, CHAIN_CODE_LENGTH},
@@ -71,7 +67,7 @@ fn create_secret(secret: &[u8]) -> SecretKey {
 ///
 /// returned vector the derived keypair as a array of 96 bytes
 #[wasm_bindgen]
-pub fn derive_keypair_hard(pair: &[u8], cc: &[u8]) -> Vec<u8> {
+pub fn ext_sr_derive_keypair_hard(pair: &[u8], cc: &[u8]) -> Vec<u8> {
 	create_from_pair(pair).secret
 		.hard_derive_mini_secret_key(Some(create_cc(cc)), &[]).0
 		.expand_to_keypair()
@@ -86,7 +82,7 @@ pub fn derive_keypair_hard(pair: &[u8], cc: &[u8]) -> Vec<u8> {
 ///
 /// returned vector the derived keypair as a array of 96 bytes
 #[wasm_bindgen]
-pub fn derive_keypair_soft(pair: &[u8], cc: &[u8]) -> Vec<u8> {
+pub fn ext_sr_derive_keypair_soft(pair: &[u8], cc: &[u8]) -> Vec<u8> {
 	create_from_pair(pair)
 		.derived_key_simple(create_cc(cc), &[]).0
 		.to_bytes()
@@ -100,7 +96,7 @@ pub fn derive_keypair_soft(pair: &[u8], cc: &[u8]) -> Vec<u8> {
 ///
 /// returned vector is the derived publicKey as a array of 32 bytes
 #[wasm_bindgen]
-pub fn derive_public_soft(public: &[u8], cc: &[u8]) -> Vec<u8> {
+pub fn ext_sr_derive_public_soft(public: &[u8], cc: &[u8]) -> Vec<u8> {
 	create_public(public)
 		.derived_key_simple(create_cc(cc), &[]).0
 		.to_bytes().to_vec()
@@ -113,7 +109,7 @@ pub fn derive_public_soft(public: &[u8], cc: &[u8]) -> Vec<u8> {
 /// returned vector is the concatenation of first the private key (64 bytes)
 /// followed by the public key (32) bytes.
 #[wasm_bindgen]
-pub fn keypair_from_seed(seed: &[u8]) -> Vec<u8> {
+pub fn ext_sr_from_seed(seed: &[u8]) -> Vec<u8> {
 	create_from_seed(seed)
 		.to_bytes()
 		.to_vec()
@@ -130,7 +126,7 @@ pub fn keypair_from_seed(seed: &[u8]) -> Vec<u8> {
 ///
 /// * returned vector is the signature consisting of 64 bytes.
 #[wasm_bindgen]
-pub fn sign(public: &[u8], secret: &[u8], message: &[u8]) -> Vec<u8> {
+pub fn ext_sr_sign(public: &[u8], secret: &[u8], message: &[u8]) -> Vec<u8> {
 	create_secret(secret)
 		.sign_simple(SIGNING_CTX, message, &create_public(public))
 		.to_bytes()
@@ -143,7 +139,7 @@ pub fn sign(public: &[u8], secret: &[u8], message: &[u8]) -> Vec<u8> {
 /// * message: Arbitrary length UIntArray
 /// * pubkey: UIntArray with 32 element
 #[wasm_bindgen]
-pub fn verify(signature: &[u8], message: &[u8], public: &[u8]) -> bool {
+pub fn ext_sr_verify(signature: &[u8], message: &[u8], public: &[u8]) -> bool {
 	let signature = match Signature::from_bytes(signature) {
 		Ok(signature) => signature,
 		Err(_) => return false
@@ -169,7 +165,7 @@ pub mod tests {
 	#[test]
 	fn can_create_keypair() {
 		let seed = generate_random_seed();
-		let keypair = keypair_from_seed(seed.as_slice());
+		let keypair = ext_sr_from_seed(seed.as_slice());
 
 		assert!(keypair.len() == KEYPAIR_LENGTH);
 	}
@@ -178,7 +174,7 @@ pub mod tests {
 	fn creates_pair_from_known() {
 		let seed = hex!("fac7959dbfe72f052e5a0c3c8d6530f202b02fd8f9f5ca3580ec8deb7797479e");
 		let expected = hex!("46ebddef8cd9bb167dc30878d7113b7e168e6f0646beffd77d69d39bad76b47a");
-		let keypair = keypair_from_seed(&seed);
+		let keypair = ext_sr_from_seed(&seed);
 		let public = &keypair[SECRET_KEY_LENGTH..KEYPAIR_LENGTH];
 
 		assert_eq!(public, expected);
@@ -187,11 +183,11 @@ pub mod tests {
 	#[test]
 	fn can_sign_message() {
 		let seed = generate_random_seed();
-		let keypair = keypair_from_seed(seed.as_slice());
+		let keypair = ext_sr_from_seed(seed.as_slice());
 		let private = &keypair[0..SECRET_KEY_LENGTH];
 		let public = &keypair[SECRET_KEY_LENGTH..KEYPAIR_LENGTH];
 		let message = b"this is a message";
-		let signature = sign(public, private, message);
+		let signature = ext_sr_sign(public, private, message);
 
 		assert!(signature.len() == SIGNATURE_LENGTH);
 	}
@@ -199,13 +195,14 @@ pub mod tests {
 	#[test]
 	fn can_verify_message() {
 		let seed = generate_random_seed();
-		let keypair = keypair_from_seed(seed.as_slice());
+		let keypair = ext_sr_from_seed(seed.as_slice());
 		let private = &keypair[0..SECRET_KEY_LENGTH];
 		let public = &keypair[SECRET_KEY_LENGTH..KEYPAIR_LENGTH];
 		let message = b"this is a message";
-		let signature = sign(public, private, message);
+		let signature = ext_sr_sign(public, private, message);
+		let is_valid = ext_sr_verify(&signature[..], message, public);
 
-		assert!(verify(&signature[..], message, public));
+		assert!(is_valid);
 	}
 
 	#[test]
@@ -213,8 +210,8 @@ pub mod tests {
 		let cc = hex!("0c666f6f00000000000000000000000000000000000000000000000000000000"); // foo
 		let seed = hex!("fac7959dbfe72f052e5a0c3c8d6530f202b02fd8f9f5ca3580ec8deb7797479e");
 		let expected = hex!("40b9675df90efa6069ff623b0fdfcf706cd47ca7452a5056c7ad58194d23440a");
-		let keypair = keypair_from_seed(&seed);
-		let derived = derive_keypair_soft(&keypair, &cc);
+		let keypair = ext_sr_from_seed(&seed);
+		let derived = ext_sr_derive_keypair_soft(&keypair, &cc);
 		let public = &derived[SECRET_KEY_LENGTH..KEYPAIR_LENGTH];
 
 		assert_eq!(public, expected);
@@ -225,8 +222,9 @@ pub mod tests {
 		let cc = hex!("0c666f6f00000000000000000000000000000000000000000000000000000000"); // foo
 		let public = hex!("46ebddef8cd9bb167dc30878d7113b7e168e6f0646beffd77d69d39bad76b47a");
 		let expected = hex!("40b9675df90efa6069ff623b0fdfcf706cd47ca7452a5056c7ad58194d23440a");
+		let derived = ext_sr_derive_public_soft(&public, &cc);
 
-		assert_eq!(derive_public_soft(&public, &cc), expected);
+		assert_eq!(derived, expected);
 	}
 
 	#[test]
@@ -234,8 +232,8 @@ pub mod tests {
 		let cc = hex!("14416c6963650000000000000000000000000000000000000000000000000000"); // Alice
 		let seed = hex!("fac7959dbfe72f052e5a0c3c8d6530f202b02fd8f9f5ca3580ec8deb7797479e");
 		let expected = hex!("d43593c715fdd31c61141abd04a99fd6822c8558854ccde39a5684e7a56da27d");
-		let keypair = keypair_from_seed(&seed);
-		let derived = derive_keypair_hard(&keypair, &cc);
+		let keypair = ext_sr_from_seed(&seed);
+		let derived = ext_sr_derive_keypair_hard(&keypair, &cc);
 		let public = &derived[SECRET_KEY_LENGTH..KEYPAIR_LENGTH];
 
 		assert_eq!(public, expected);