+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 0.3.1
+
+- Add wasm-dalek-ed25519
+- Add keccak256
+- Try secp256k1 (bloating code, not used atm)
+
+# 0.2.1
+
+- Add basic crypto functions, creation of wasm-crypto package
+- blake2, bip39, sha2, xxhash
+
 # 0.1.1
 
 - Initial release